requirements.txt

kafka-python
uvicorn
googletrans==3.1.0a0
nltk
ormar
sqlalchemy
asyncpg==0.27.0
fastapi==0.89.1
ormar==0.12.1
psycopg2-binary==2.9.5
uvicorn==0.20.0